A/B 다른 코드

+++ baekjoon/1008-1.c
@@ -2,9 +2,9 @@
 
 int main()
 {
-    int A, B;
+    double A, B;
 
-    scanf("%d %d", &A, &B);
-    printf("%.9f", (double) A / B);
+    scanf("%lf %lf", &A, &B);
+    printf("%.9f", A / B);
     return 0;
 }
